Switch to using Cmake's FetchContent
This is the only way to make things sane when we later need to integrate with Synthizer unless we want to take a hard dependency on a proper
package manager, which would in turn cause a huge amount of downstream pain.  So, less optimal but will get the job done path it is.